Update time_util to be working on linux
Bench: 18866923

+++ src/time_util.c
@@ -22,7 +22,8 @@ uint64_t get_time_ms(void) {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
 
-    return ts.tv_sec * 1000ULL + ts.tv_nsec / 1000000ULL;
+    return (uint64_t)(((uint64_t)ts.tv_sec * 1000ULL) +
+                      ((uint64_t)ts.tv_nsec / 1000000ULL));
 
 #endif
 }
